Modcommands: Update
Add optional reason parameter for mods that dont want to add to the spreadsheet later and would rather get it out of the way. Also fix a syntax error in the code.
Also add a duration check to the code to make sure its a number
Add a check for the new timeout function for future handling of any issues in case it malfunctions